eDVBFrontendParametersTerrestrial / frequency must be int

+++ lib/python/Components/Converter/ChannelNumbers.py
@@ -60,9 +60,9 @@ class ChannelNumbers:
 		descr = self.getTunerDescription(nim)
 		if "Europe" in descr and "DVB-T" in descr:
 			if 5 <= channel <= 12:
-				return (177500 + 7000 * (channel - 5)) * 1000
+				return int((177500 + 7000 * (channel - 5)) * 1000)
 			elif 21 <= channel <= 69:
-				return (474000 + 8000 * (channel - 21)) * 1000
+				return int((474000 + 8000 * (channel - 21)) * 1000)
 		return 474000000
